group_authorization == loose gives permission to all groups
In strict group authorization (AND operator between all host authorizations
tests) bails out with a return false if ANY host isn't authorized. This is
correct.

In loose group authorization (OR operator between all host authrozations tests)
bails out with a return true if ANY host is authorized. This is also correct.

But having a default return true at the end if not bailed out seems to be a
good idea if returning false in all other cases before. This is ok for strict
autorization.

But default returning true, if all other cases has returned... well... true
before might be an interesting case, that well... never could have worked... If
not returned true before, return true now.

Signed-off-by: Max Sikström <msikstrom@op5.com>