authorization: Change handling of empty groups
Before, a user was always authorized for a group if the group had no
members - and, after the parent commit, we also had to be in strict
mode for this to be a problem.

It's a tricky case - we're not authorized for the group, but we're also
not /not/ authorized. However, my motivation is two-fold:

A customer let us know that they were unhappy with the current behavior,
and could motivate the unhappiness properly: when you're preparing
configuration for new limited users, you don't want your old, unrelated
limited users to see the new (partial) configuration.

Additionally, it intuitively makes sense that the "strict" mode only
lets you see groups where you're contact for at least one member - the
old behaviour was that it only let you see groups where you're not
unauthorized for anything.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>