Filter out custom_variables that shouldn't be exporeted using prefix
To be able to use custom variables to store certain configuration parameters to
services and host checks, such as service account information, those variables
needs not to be exported to any GUI, but just be kept inside of nagios.

That means, those variables needs to be masked in the inteface between nagios
and the GUI, which for most cases is livestatus.

This patch adds the configuration line hidden_custom_var_prefix=PREFIX to the
broker_module line in nagios.cfg, which makes all custom variables with that
particular prefix hidden in the livestatus interface.

Should resolve a big part of #7894

Signed-off-by: Max Sikström <msikstrom@op5.com>