spec: Fix regexes
5.7 (or so) contained the livestatus line
    broker_module=/opt/monitor/op5/livestatus/livestatus.o
We then started adding more and more options to the line, and part of
the way through, we moved the module.

Now, as we moved the module, all the regexes' match part stopped
matching the old path, which meant that none of the migration steps
worked anymore.

Instead, migrate the path first, and then start doing all the regular
migrations, one by one.

Also, always go through all the migrations (except the path) - we don't
need to complicate the first regex, when we know the pnp_path will be
added on the very next line.

Change-Id: I2be5b718ee512dd72a7892c04f1ea36b47b120ad
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>