AUTHORS: Remove more op5 references
The quirky phrasing in the AUTHORS file is intended to say "do credit mk
with creating this, but don't go to him if you find bugs, because we
changed it and probably broke it in the process".

But this particular fork (which will probably become the upstream for
the op5 one, rather than the historical inverse) is maintained by
naemon, as the readme suggests.

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>