configure: Rename package to naemon-livestatus
Since this is not mk-livestatus, we should pick a better name. Just
'livestatus' is no good, since that has become a collective name for all
different forks and implementations of the original. And since
compatibility with Naemon is what we really strive for here, it feels
sane to make the name reflect that.

Change-Id: I83f28f44e12fb9661b735c226fefe3674f6ca73f
Signed-off-by: Anton Lofgren <alofgren@op5.com>