tests: Trigger insane regex find bug
This patch triggers (in a test case) a bug in the unicode  regex filter
implementation that causes filtering to work rather peculiarly. In
essence, because of how C++ treats anonymous objects, the current
implementation causes an internal pointer in the ICU library to point at
garbage (often some offset into the filter string), which results in
very unintuitive "matches".

Change-Id: I6fb6a2b7a1e1d63a5512c6a96fadc57ba31a92a2
Signed-off-by: Anton Lofgren <alofgren@op5.com>