filters: Don't pass anonymous objects around
Anonymous objects are implicitly destroyed if not bound. This leads to
the RegexMatcher's string being different than the one we attempt to
match against (by an offset of 2 bytes, it seems).

My original test cases did not catch this since they mostly differ in
the end of the string.

This fixes a bug where the regular expression filters behaved _really_
unintuitively.

Change-Id: I20e8da885c4d8fbbf9441b7b372ccc14515e64a5
Signed-off-by: Anton Lofgren <alofgren@op5.com>