configure: Fallback for ICU if no .pc found
If we don't find a .pc for libicu, we now print a warning and use some
flags that should work (provided that libicu actually is installed).
This is _not_ the right way to do it, but at least it'll get the thing
built.

This also includes being particular about the version of libicu we
accept, since we can only cope with >= 4.2.

Change-Id: Ifb24852e12b6986a777724f67836cfafa1a6f26a
Signed-off-by: Anton Lofgren <alofgren@op5.com>