specfile: Fix regex for determining if the module is already enabled
When several modules are dlopen()'ed from the same path and then are
closed again, glibc on many systems will cause a segfault.

The old regex checked for livestatus.o, which the file was called
once upon a time. This fixes it so the regex checks for
livestatus.so, which it's called nowadays. This will prevent the
multiple-of-the-same-module problem from appearing again.

We also fix the other places, avoiding the oddness where we add
hidden_custom_var_prefix and pnp paths multiple times to the
broker_module line.

Change-Id: I71e66690b6f1fddcec723782c3e8576826672c5f
Signed-off-by: Andreas Ericsson <ae@op5.se>