filters: Don't block macro column filtering
Since macro column filtering was blocked (by using an AndingFilter),
all filtering on macro columns was completely ineffectual. This patch
removes the blocking, enabling filtering on such columns. However,
macros are still not resolved. No change here, they never were, but it's
a somewhat unintuitive behaviour which we might want to fix in a future
release.

This fixes #8435.

Change-Id: I2e48b89a5f407ea5f6b51b5d17038b028b5e0af3
Signed-off-by: Anton Lofgren <alofgren@op5.com>