Revert "init: Start threads on nebmodule_init"
This reverts commit 8f41d8469c8c0b9f846000cc2080256c34400c86.

Apparently (according to MS), this is bat country (not his words),
because we've initialized a bunch of stuff, which may lead to
interesting results should some attempt to query livestatus while
everything is not yet fully settled. I haven't looked into it further,
but it seems the gain is not worth the risk at the moment. So, revert.