module: Send termination message to main thread
Since the main thread is most likely stuck in a select() call when we're
ordered to shut down, we now open a pipe pair on initialisation for
"internal" IPC - the read thread being added to the read set in the main
thread's select() call. Currently, this pipe only serves to trigger an
interrupt in the main thread. In other words, the actual message could
be anything. This could easily be changed however, should we want to add
more messages.

This patch will make the thread terminate in an orderly fashion without
having to wait for an eternity (a.k.a 2.5 seconds). This, in turn speeds
up a lot of things, including test cases, installation, configuration
reloading and probably a bunch of other things.

Change-Id: I1d6245ec7b5759a6a0a8987d796884878ec1b022
Signed-off-by: Anton Lofgren <alofgren@op5.com>