autoconf: Handle ICU properly
This patch adds a macro to check for ICU (setting, _CFLAGS, _CXXFLAGS and
_LIBS) instead of trying to do so ourselves, when PKG_CHECK_MODULES
fails. The script has been slightly modified from its original to work better with our
current build system. It uses the icu-config program to retrieve the
proper paths, which is deprecated in favour of pkg-config (and other
techniques) according to http://userguide.icu-project.org/howtouseicu.
But, since some distibutions don't have a recent enough version of ICU
available, we need this fallback.

Of course, this deprecates the old "guess which flags to use based on
dubious heuristics" approach we used to take.

It also gets rid of the repeated "checking for ICU ... xx" printouts we
used to get when the fallback was triggered by nesting the fallbacks
instead.

Change-Id: Ibdf127f395160359cd484b4eecf7ac70f9f4a681
Signed-off-by: Anton Lofgren <alofgren@op5.com>