hosts, services: Generate should_be_scheduled based on what it meant
should_be_scheduled used to be set and paid attention to in nagios
according to if a host/service had a check interval, and didn't have
disabled checks.

As the code has been consolidated and the check is only done once, the
variable isn't needed in naemon. Since even disabled checks are now
scheduled, regardless of whether whey actually run or not, the value
wouldn't even be completely truthful if it actually was set in naemon.

Hence, define the column according to what it's supposed to mean. This
avoids breaking livestatus and scripts using livestatus, without
introducing the mess back into naemon.

Change-Id: I3edb58243fcd98855b9f62c1ffaba4231c9f5e8d
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>