Makefile.am: Set NAEMON_CFLAGS in CXXFLAGS as well
It used to be set in CPPFLAGS (C preprocessor flags), which applies to
both C and C++. However, that variable was recently renamed to
CFLAGS, which doesn't apply to C++. And the C++ code uses naemon.

This fixes builds when naemon's headers are installed somewhere where
the system's default CFLAGS doesn't include.

Change-Id: I2077c1db4a8569757b394a4d658835295ab74936
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>