Replace 'main thread' of livestatus with Naemon's I/O broker
The main thread of livestatus was really only used to poll the
unix socket for input, and then react to it by queueing the
accept()'ed socket for the first available thread to read and
process the request.

Since the introduction of the iobroker into Naemon, however,
that's not really necessary anymore, as we already do everything
the main thread used to do.

While we're refactoring, we also get rid of the atfork handlers
and the bookkeeping needed for them to work. They were pretty
intricately nested with the launching of the main thread anyway,
so a separate patch would have been convoluted and clunky, and
we don't need the pthread_atfork() handlers since we now use
workers instead of fork()'ing the main process for every single
check we run.

Gone is also the clunky ipc pipe which was previously used to
communicate between the main naemon process and the main
livestatus thread in a threadsafe matter.

Since this removes any possibility of a race between the main
thread and the naemon core thread (as they're now one and the
same), this resolves MON-7966.

Change-Id: I1fcaae3cc246a99f85b99ff54cbbc9f7c56ac9e4
Signed-off-by: Andreas Ericsson <ae@op5.se>