Remove limit on number of concurrent requests
This patch removes an artificial limit on the number of concurrent
requests we can handle. It does this by getting rid of the naïve
thread-pool implementation and replacing it with plain old detached
threads. In addition to avoiding resource starvation, this has the
benefit of cleaning up the code: For example, by getting rid of the
ClientQueue, and the thread synchronization bits.

This fixes MON-8803.

Change-Id: I2e8212b22073407e7ff80e087fb0d5e7f445f6bb
Signed-off-by: Anton Lofgren <alofgren@op5.com>