Add the "notified_on" columns for hosts and services
This makes it relatively straightforward to check if all markers
are set that naemon uses to track if a notification has been sent
or not, which is insanely nifty when writing tests to track
notification behaviour in various scenarios, such as a distributed
setup.

This is part of creating automated tests for MON-8720.

Change-Id: Ia4fdd3c87bad51d0ae4db7d7efe402ea3c89fb95
Signed-off-by: Andreas Ericsson <ae@op5.se>