Add mutex to protect the shared log file
Since we now spawn threads detached (be8fda5d82c4), we no longer wait
for client threads to terminate before we close the shared log file,
which leads to a potential race. The race consists of the main thread
closing the log file on exit, and a client thread trying to log to the
(non-invalidated) fd before the main thread has fully terminated.

In most cases, this will simply result in no logging being done, but
can also lead to the fd pointing at some other file in the process. This
latter case could have disastrous consequences, in which Naemon hangs
indefinitely on reload.

We fix this by adding a mutex for the shared resource.

Of course, this should've always been protected to avoid multiple
concurrent writes, but it's probably the case that we rarely write
enough bytes often enough to trigger any visible symptoms.

This fixes MON-9055.

Change-Id: I5cdb6e8f3fd6042a71e44d7ecd8bd5a8c0e50d82
Signed-off-by: Anton Lofgren <alofgren@op5.com>