Keep track of threads again
Detaching the threads leads to issues on restart, since we're liable to
touch global naemon state that has been freed in our client threads.
This patch makes sure to keep track of how many threads we've created,
and joining with them before returning from deinit.

This does *not* mean that we reintroduce the thread pool. Rather, we
create and join threads dynamically instead.

This fixes MON-9123.

Signed-off-by: Anton Lofgren <alofgren@op5.com>