Actually set default config install path
The mainconf variable is defined by the naemon .pc file, but it sems
we've never actually bothered to read it before - instead hoping that
everyone overrides the naemon-config-dir when running ./configure.

This patch makes sure to set the variable so that doing a pure
./configure actually works as expected, and doesn't break the build by
trying to install livestatus.cfg as ./livestatus.cfg.

Signed-off-by: Anton Lofgren <alofgren@op5.com>