build system: Enable silent rules if available
This project produces somewhat nasty warnings when building,
but they're mostly obscured by the (extremely noisy) autoconf
build output. Shut most of it up so warnings stand out and
can be fixed more easily.

This patch is janitorial in preparation for the improvements.
As such it lacks relation to a specific issue.

Signed-off-by: Andreas Ericsson <ae@op5.se>