Make broker module options configurable
This patch allows users to specify default options for the broker module
configuration, which comes in handy for distributors that are not
satisfied with the defaults - for example for setting a custom socket
path.

The defaults are still the same as they were prior to this change,
provided that the --with-broker-module-options is not used with
./configure.

In addition, this patch updates the op5 dist spec to use this instead of
overwriting the default configuration file.

Signed-off-by: Anton Lofgren <alofgren@op5.com>