Servicelist should filter on what it shows
The filtering on host only makes sense in the case the servicelist is
displayed with hosts, which is in the servicegroup table.

But in the case of hosts, with the services list, it doesn't make sense
at all. In fact, it only outputs the service description.

Thus, filter on what it displays, which makes sense. Everything else
must be a bug.

Resolves MON-9480

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>