send error response on too large results
so far livestatus would quietly end a query if it hits the maximum output size without
the client being able to notice. Since it is better to get no data and an error instead
of wrong or incomplete data and not even knowing it we better bail out with an error.

Signed-off-by: Sven Nierlein <sven@nierlein.de>