Add support for TCP sockets
This change adds support for TCP sockets, in addition to UNIX sockets to
Livestatus. The idea is to be able to connect to Naemon/Livestatus over
TCP without having to use an additional daemon, like xinetd.

Add event broker option "inet_addr" that can be used instead of
providing the UNIX socket path. It takes an "IPv4:port" tuple as
argument, where an address of 0.0.0.0 means to listen on all available
interfaces. Example config:

  broker_module=/usr/lib64/livestatus.so inet_addr=0.0.0.0:32768

UNIX socket is still specified as before, with a lone path argument:

  broker_module=/usr/lib64/livestatus.so /usr/local/nagios/var/rw/live

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>