tests: update connection test in ls-query
Update ls-query so that when test suite runs
`ls-query.py --test-connect`, it tries to run an actual livestatus query
('GET status') instead of just checking if the tcp socket is up.
This should make tests more robust, so that they don't think Naemon with
broker modules are up and running and start sending external commands,
before modules are loaded properly.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>