tests: checks long plugin output handling in submitted commands
This scenario submits a "long" output (multiline, newline separated) and
checks if the first line is properly extracted as `plugin_output` while
the remaining ones should go to `long_plugin_output`.

This is in fact testing the 'unescape' of newlines (and other characters)
from livestatus `COMMAND`s.

Preliminary work for #387

Signed-off-by: Ludovic LANGE <llange@users.noreply.github.com>