add service parents host_name
according to our docs, the service parents can be:

```
 Valid option is either a single service description from the same host or a comma separated list of host_name,servicedescription tupel.
```
which means you could also define service parent links to different hosts which then means we have to list hostname and description in the
parent list. Otherwise we cannot know which host the service is on and assuming it is the same host could be wrong.