fix too low thread stacksize
thread stack size cannot be lower than PTHREAD_STACK_MIN. On some systems, ex.: debian 11 on aarch64, the lower limit is 1M. This results
in lots of `livestatus: Error: Cannot set thread stack size to 65536` entries in the logfile.
So lets check if the value is at least PTHREAD_STACK_MIN to avoid this error message.