Restrict "Keep track of threads again" to glibc
The fix introduced in 1bd7a3b385a887dc3537480095cb6ce30c2dee76
will only compile with glibc, as it uses additions to pthreads
specific to the GNU C Library.

This fix allows in certain conditions to solve a threads issue
when the livestatus broker restarts.

As such, it seems currently best to keep it for Linux systems,
and skip it elsewhere to allow livestatus to work everywhere else.