logcache: check connection before building logs result
the logcache uses a lock to only server one request at a time. Pending
requests will be queued and served in order, even if the client has
disconnected meanwhile. The log request would be build in this case
and then livestatus detects the broken pipe once it sends out the first
byte.
Better check the client connection once before building the result.